MD: Fix all leaks, creating 'check if its a valid file', refactory every malloc, algo done but need improve

+++ libft/ft_strlen.c
@@ -12,8 +12,8 @@
 
 #include "libft.h"
 
-size_t ft_strlen(const char *str) {
-  size_t i;
+int ft_strlen(const char *str) {
+  int i;
 
   i = 0;
   while (str[i])
